gotta fix malloc problem when a big command is passed and ls args peoblem at 12-18

+++ strfucs2.c
@@ -33,16 +33,16 @@ return (dest);
 
 int _strcmp(char *s1, char *s2)
 {
-int ans = 0;
-while (*s1 && *s2)
-{
-if (*s1 != *s2)
-	return (*s1 - *s2);
-s1++;
-s2++;
-}
+	int ans = 0;
+	while (*s1 && *s2)
+	{
+		if (*s1 != *s2)
+			return (int)(*s1 - *s2);
+		s1++;
+		s2++;
+	}
 
-return (*s1 - *s2);
+	return (int)(*s1 - *s2);
 }
